refactor(spec-047): relocate Bucket 1 legacy mount/update into V1-owned code
Move the control-specific mount/update logic for the remaining Path-B
delegator handlers out of the shared Reconciler.Mount.cs/Reconciler.Update.cs
files into V1-owned lifecycle/handler classes, then delete the legacy bodies.
Pure verbatim relocation — no behavior change.

- G1 CheckBox + Expander: folded Mount/Update into the handler classes
- G2 ListView + GridView: handlers own their realize/recycle logic
- G3 NavigationHost: extracted NavigationHostLifecycle.cs
- G4 LazyStack: extracted LazyStackLifecycle.cs (deduped BuildListState)
- G5 Templated List/Grid/FlipView: extracted TemplatedListLifecycle.cs;
  TemplatedTreeView<T> moved verbatim into Reconciler.TemplatedTree.cs as
  partial-class Reconciler members (kept coupled to UnmountRecursive + the
  _typedTreeListControls CWT rather than a separate lifecycle class)

Shared helpers used by KeyedItemsBinding.cs (HandleTemplatedContainerContentChanging,
RefreshRealizedContainers, ApplyMoveAnimations) widened to internal and kept in
Reconciler. Reflection selftests repointed to the relocated methods; added
NavHostCacheMode selftest coverage.

Net: -1595 lines from the two shared reconciler files. Unit (9132 passed) and
selftest (1006 passed) suites green; 0 failures.
