+++ src/Reactor/Core/Reconciler.Mount.cs
@@ -238,135 +238,11 @@ public sealed partial class Reconciler
     }
 
 
-    internal WinUI.CheckBox MountCheckBox(CheckBoxElement cb)
-    {
-        var checkBox = new WinUI.CheckBox { Content = cb.Label };
-        if (cb.IsThreeState)
-        {
-            checkBox.IsThreeState = true;
-            checkBox.IsChecked = cb.CheckedState;
-        }
-        else
-        {
-            checkBox.IsChecked = cb.IsChecked;
-        }
-        SetElementTag(checkBox, cb);
-        if (cb.OnIsCheckedChanged is not null || cb.OnCheckedStateChanged is not null)
-        {
-            checkBox.Checked += (s, _) =>
-            {
-                var c = (UIElement)s!;
-                if (ChangeEchoSuppressor.ShouldSuppress(c)) return;
-                var el = GetElementTag(c) as CheckBoxElement;
-                el?.OnIsCheckedChanged?.Invoke(true);
-                el?.OnCheckedStateChanged?.Invoke(true);
-            };
-            checkBox.Unchecked += (s, _) =>
-            {
-                var c = (UIElement)s!;
-                if (ChangeEchoSuppressor.ShouldSuppress(c)) return;
-                var el = GetElementTag(c) as CheckBoxElement;
-                el?.OnIsCheckedChanged?.Invoke(false);
-                el?.OnCheckedStateChanged?.Invoke(false);
-            };
-            checkBox.Indeterminate += (s, _) =>
-            {
-                var c = (UIElement)s!;
-                if (ChangeEchoSuppressor.ShouldSuppress(c)) return;
-                var el = GetElementTag(c) as CheckBoxElement;
-                el?.OnCheckedStateChanged?.Invoke(null);
-            };
-        }
-        ApplySetters(cb.Setters, checkBox);
-        return checkBox;
-    }
 
 
 
 
-    internal WinUI.Expander MountExpander(ExpanderElement exp, Action requestRerender)
-    {
-        var expander = new WinUI.Expander
-        {
-            IsExpanded = exp.IsExpanded,
-            ExpandDirection = exp.ExpandDirection,
-        };
-        // Element header wins over the string slot (matches the spec
-        // "HeaderTemplate" slot semantics — strings are still supported as
-        // the default header content).
-        if (exp.HeaderTemplate is not null)
-            expander.Header = Mount(exp.HeaderTemplate, requestRerender);
-        else
-            expander.Header = exp.Header;
-        if (exp.ContentTransitions is not null) expander.ContentTransitions = exp.ContentTransitions;
-        expander.Content = Mount(exp.Content, requestRerender);
-        SetElementTag(expander, exp);
-        if (exp.OnIsExpandedChanged is not null)
-        {
-            expander.Expanding += (s, _) => (GetElementTag((UIElement)s!) as ExpanderElement)?.OnIsExpandedChanged?.Invoke(true);
-            expander.Collapsed += (s, _) => (GetElementTag((UIElement)s!) as ExpanderElement)?.OnIsExpandedChanged?.Invoke(false);
-        }
-        ApplySetters(exp.Setters, expander);
-        return expander;
-    }
 
-
-
-    internal WinUI.Grid MountNavigationHost(NavigationHostElement element, Action requestRerender)
-    {
-        var grid = new WinUI.Grid();
-        var handle = (Navigation.INavigationHandle)element.NavigationHandle;
-        var routeMap = element.RouteMap;
-        var currentRoute = handle.CurrentRoute;
-
-        // Resolve and mount the initial route's element
-        var childElement = routeMap(currentRoute);
-        var childControl = Mount(childElement, requestRerender);
-        if (childControl is not null)
-            grid.Children.Add(childControl);
-
-        // Track state for update/unmount
-        var node = new NavigationHostNode
-        {
-            Handle = handle,
-            LastRenderedRoute = currentRoute,
-            CurrentChildElement = childElement,
-            CurrentChildControl = childControl,
-            RouteMap = routeMap,
-            RequestRerender = requestRerender,
-            HostTransition = element.Transition,
-            CacheMode = element.CacheMode,
-        };
-
-        // Create page cache when caching is enabled
-        if (element.CacheMode != Navigation.NavigationCacheMode.Disabled)
-        {
-            node.Cache = new Navigation.NavigationCache(
-                element.CacheSize, evicted => Unmount(evicted));
-        }
-
-        // Subscribe to route changes so NavigationHost updates even if an intermediate
-        // component's ShouldUpdate blocks the re-render propagation.
-        void onRouteChanged() => requestRerender();
-        handle.RouteChanged += onRouteChanged;
-        node.RouteChangedHandler = onRouteChanged;
-
-        // Wire lifecycle guard: invokes onNavigatingFrom callbacks from the current
-        // page's component tree before the stack mutation. Records the navigation mode
-        // and previous route for post-swap onNavigatedTo/onNavigatedFrom invocation.
-        handle.LifecycleGuard = ctx =>
-        {
-            InvokeNavigatingFrom(node.CurrentChildControl, ctx);
-            if (!ctx.IsCancelled)
-            {
-                node.PendingNavigationMode = ctx.Mode;
-                node.PendingPreviousRoute = ctx.Route;
-            }
-        };
-
-        _navigationHostNodes[grid] = node;
-        return grid;
-    }
 
     private static WinUI.NavigationViewItem CreateNavItem(NavigationViewItemData data)
     {
@@ -489,147 +365,6 @@ public sealed partial class Reconciler
         return btn;
     }
 
-    // Internal visibility so V1Protocol.Handlers.ListViewHandler can delegate
-    // to the shared mount body (Path B from spec 047 §14 Phase 1 task 1.15:
-    // delegate + shape declaration rather than full reimplementation, so the
-    // V1 ON and V1 OFF paths execute identical mount bodies).
-    internal WinUI.ListView MountListView(ListViewElement lv, Action requestRerender)
-    {
-        var listView = new WinUI.ListView
-        {
-            SelectionMode = lv.SelectionMode,
-            IsItemClickEnabled = lv.OnItemClick is not null,
-            IncrementalLoadingTrigger = lv.IncrementalLoadingTrigger,
-        };
-        if (lv.Header is not null) listView.Header = lv.Header;
-        if (lv.ItemContainerStyle is not null) listView.ItemContainerStyle = lv.ItemContainerStyle;
-
-        SetElementTag(listView, lv);
-
-        // DataTemplate with a ContentControl shell — we populate its Content on demand
-        listView.ItemTemplate = SharedContentControlTemplate.Value;
-
-        listView.ContainerContentChanging += (sender, args) =>
-        {
-            if (args.InRecycleQueue)
-            {
-                if (args.ItemContainer.ContentTemplateRoot is ContentControl oldCc)
-                {
-                    if (oldCc.Content is UIElement oldCtrl)
-                        UnmountChild(oldCtrl);
-                    oldCc.Content = null;
-                }
-                return;
-            }
-
-            args.Handled = true;
-            var items = (GetElementTag((UIElement)sender!) as ListViewElement)?.Items;
-            if (items is not null && args.ItemIndex >= 0 && args.ItemIndex < items.Length
-                && args.ItemContainer.ContentTemplateRoot is ContentControl cc)
-            {
-                var ctrl = Mount(items[args.ItemIndex], requestRerender);
-                cc.Content = ctrl;
-            }
-        };
-
-        // Subscribe unconditionally so OnSelectionChanged (multi-select snapshot)
-        // and OnSelectedIndexChanged (single focused index) both pick up
-        // handlers attached on a later record-with without re-subscribing.
-        listView.SelectionChanged += (s, _) =>
-        {
-            var l = (WinUI.ListView)s!;
-            if (GetElementTag(l) is not ListViewElement el) return;
-            el.OnSelectedIndexChanged?.Invoke(l.SelectedIndex);
-            if (el.OnSelectionChanged is { } h)
-            {
-                // SelectedItems is List<object> of int — copy into a typed snapshot.
-                var snapshot = new List<int>(l.SelectedItems.Count);
-                foreach (var item in l.SelectedItems)
-                    if (item is int i) snapshot.Add(i);
-                h(snapshot);
-            }
-        };
-        if (lv.OnItemClick is not null)
-            listView.ItemClick += (s, args) =>
-            {
-                var l = (WinUI.ListView)s!;
-                if (args.ClickedItem is int idx)
-                    (GetElementTag(l) as ListViewElement)?.OnItemClick?.Invoke(idx);
-            };
-
-        // Set ItemsSource LAST — triggers container creation which needs the handler above
-        listView.ItemsSource = Enumerable.Range(0, lv.Items.Length).ToList();
-
-        if (lv.SelectedIndex >= 0) listView.SelectedIndex = lv.SelectedIndex;
-        ApplySetters(lv.Setters, listView);
-        return listView;
-    }
-
-    internal WinUI.GridView MountGridView(GridViewElement gv, Action requestRerender)
-    {
-        var gridView = new WinUI.GridView
-        {
-            SelectionMode = gv.SelectionMode,
-            IsItemClickEnabled = gv.OnItemClick is not null,
-            IncrementalLoadingTrigger = gv.IncrementalLoadingTrigger,
-        };
-        if (gv.Header is not null) gridView.Header = gv.Header;
-        if (gv.ItemContainerStyle is not null) gridView.ItemContainerStyle = gv.ItemContainerStyle;
-
-        SetElementTag(gridView, gv);
-
-        gridView.ItemTemplate = SharedContentControlTemplate.Value;
-
-        gridView.ContainerContentChanging += (sender, args) =>
-        {
-            if (args.InRecycleQueue)
-            {
-                if (args.ItemContainer.ContentTemplateRoot is ContentControl oldCc)
-                {
-                    if (oldCc.Content is UIElement oldCtrl)
-                        UnmountChild(oldCtrl);
-                    oldCc.Content = null;
-                }
-                return;
-            }
-
-            args.Handled = true;
-            var items = (GetElementTag((UIElement)sender!) as GridViewElement)?.Items;
-            if (items is not null && args.ItemIndex >= 0 && args.ItemIndex < items.Length
-                && args.ItemContainer.ContentTemplateRoot is ContentControl cc)
-            {
-                var ctrl = Mount(items[args.ItemIndex], requestRerender);
-                cc.Content = ctrl;
-            }
-        };
-
-        gridView.SelectionChanged += (s, _) =>
-        {
-            var g = (WinUI.GridView)s!;
-            if (GetElementTag(g) is not GridViewElement el) return;
-            el.OnSelectedIndexChanged?.Invoke(g.SelectedIndex);
-            if (el.OnSelectionChanged is { } h)
-            {
-                var snapshot = new List<int>(g.SelectedItems.Count);
-                foreach (var item in g.SelectedItems)
-                    if (item is int i) snapshot.Add(i);
-                h(snapshot);
-            }
-        };
-        if (gv.OnItemClick is not null)
-            gridView.ItemClick += (s, args) =>
-            {
-                var g = (WinUI.GridView)s!;
-                if (args.ClickedItem is int idx)
-                    (GetElementTag(g) as GridViewElement)?.OnItemClick?.Invoke(idx);
-            };
-
-        gridView.ItemsSource = Enumerable.Range(0, gv.Items.Length).ToList();
-
-        if (gv.SelectedIndex >= 0) gridView.SelectedIndex = gv.SelectedIndex;
-        ApplySetters(gv.Setters, gridView);
-        return gridView;
-    }
 
     // Legacy TreeViewNodeData.ContentElement reads — the property is [Obsolete]
     // in favor of typed TreeView<T> (issue #447) but the path stays functional
@@ -678,199 +413,16 @@ public sealed partial class Reconciler
     }
 
     // ── Typed, data-driven TreeView<T> ───────────────────────────────────
-
-    /// <summary>
-    /// Mounts a typed <see cref="TemplatedTreeViewElementBase"/>. Builds a WinUI
-    /// node-mode <c>TreeView</c> whose <c>ItemTemplate</c> is an empty
-    /// <c>ContentControl</c> shell; each node's view is mounted imperatively
-    /// into the realized container on demand (see
-    /// <see cref="OnTypedTreeContainerContentChanging"/>) — the same
-    /// realize/recycle hosting as the typed <c>ListView</c>, which keeps
-    /// expand/collapse robust under container recycling. <c>node.Content</c>
-    /// holds the developer's data item.
-    /// </summary>
-    internal WinUI.TreeView MountTemplatedTreeView(TemplatedTreeViewElementBase el, Action requestRerender)
-    {
-        var treeView = new WinUI.TreeView
-        {
-            SelectionMode = el.GetSelectionMode(),
-            CanDragItems = el.GetCanDragItems(),
-            AllowDrop = el.GetAllowDrop(),
-            CanReorderItems = el.GetCanReorderItems(),
-            ItemTemplate = SharedContentControlTemplate.Value,
-        };
-
-        foreach (var root in el.GetRoots())
-            treeView.RootNodes.Add(BuildTemplatedTreeNode(el, root));
-
-        SetElementTag(treeView, el);
-
-        // Trampolines resolve the live element + the data item (node.Content)
-        // on dispatch, so they're wired unconditionally and never need
-        // re-subscribing on Update (a no-op when the user supplied no callback).
-        treeView.ItemInvoked += TemplatedTreeItemInvoked;
-        treeView.Expanding += TemplatedTreeExpanding;
-
-        // Hook the internal TreeViewList ("ListControl") ContainerContentChanging
-        // so node views are mounted into their realized containers. The
-        // ListControl only exists once the template applies (after the control
-        // loads in-tree), so we subscribe on Loaded and populate the
-        // already-realized initial containers there. Loaded also re-attaches
-        // after an Unloaded/Loaded cycle; the attach is idempotent.
-        treeView.Loaded += (s, _) => AttachTypedTreeHosting((WinUI.TreeView)s!, requestRerender);
-
-        el.ApplyControlSetters(treeView);
-        return treeView;
-    }
-
-    /// <summary>
-    /// Recursively materializes a <see cref="WinUI.TreeViewNode"/> for
-    /// <paramref name="item"/>. <c>node.Content</c> holds the data item; the
-    /// view itself is mounted lazily per realized container, so nothing is
-    /// mounted here.
-    /// </summary>
-    private static WinUI.TreeViewNode BuildTemplatedTreeNode(TemplatedTreeViewElementBase el, object item)
-    {
-        var node = new WinUI.TreeViewNode { Content = item, IsExpanded = el.GetIsExpanded(item) };
-        var children = el.GetChildren(item);
-        if (children is not null)
-            foreach (var child in children)
-                node.Children.Add(BuildTemplatedTreeNode(el, child));
-        return node;
-    }
-
-    /// <summary>
-    /// Subscribes the typed TreeView's internal list to ContainerContentChanging
-    /// (once) and populates any containers that realized before the subscription.
-    /// Idempotent — presence in <see cref="_typedTreeListControls"/> marks
-    /// "already subscribed", so it's safe to call on every Loaded.
-    /// </summary>
-    private void AttachTypedTreeHosting(WinUI.TreeView treeView, Action requestRerender)
-    {
-        if (_typedTreeListControls.TryGetValue(treeView, out _)) return; // already subscribed
-
-        var list = FindDescendantListView(treeView);
-        if (list is null) return; // template not applied yet — a later Loaded will retry
-
-        _typedTreeListControls.Add(treeView, list); // mark subscribed + cache for Update
-        list.ContainerContentChanging += (_, args) =>
-            OnTypedTreeContainerContentChanging(treeView, args, requestRerender);
-
-        // Host the containers that realized before we subscribed — their
-        // ContainerContentChanging already fired and won't fire again. They may
-        // not be ready yet (their ContentTemplateRoot is generated a layout pass
-        // later — observed under NativeAOT, where the realized container is even
-        // still the base ListViewItem at Loaded time), so re-attempt on
-        // LayoutUpdated until every realized container is hosted, then detach.
-        // Everything realized AFTER this point flows through CCC. The pass count
-        // is bounded so the handler always detaches (no dangling subscription),
-        // and CCC still covers anything not hosted by then.
-        if (!PopulateRealizedTreeContainers(treeView, list, requestRerender))
-        {
-            int passesLeft = 8;
-            EventHandler<object>? onLayout = null;
-            onLayout = (_, _) =>
-            {
-                if (PopulateRealizedTreeContainers(treeView, list, requestRerender) || --passesLeft <= 0)
-                    list.LayoutUpdated -= onLayout;
-            };
-            list.LayoutUpdated += onLayout;
-        }
-    }
-
-    /// <summary>
-    /// Hosts every currently-realized container that's ready and not yet hosted.
-    /// Returns true when no realized container remains unhosted (so the caller
-    /// can stop re-attempting). Virtualized-out indices (null container) are not
-    /// counted — ContainerContentChanging hosts them when they realize.
-    /// </summary>
-    private bool PopulateRealizedTreeContainers(WinUI.TreeView treeView, WinUI.ListView list, Action requestRerender)
-    {
-        bool complete = true;
-        for (int i = 0; i < list.Items.Count; i++)
-        {
-            // The container is a ListViewItem/TreeViewItem (both ContentControl).
-            // Don't filter on TreeViewItem — under AOT it can still be the base
-            // ListViewItem when first realized.
-            if (list.ContainerFromIndex(i) is not ContentControl container) continue;
-            if (container.ContentTemplateRoot is null) { complete = false; continue; } // not ready yet
-            PopulateTypedTreeContainer(treeView, container, list.Items[i], requestRerender);
-        }
-        return complete;
-    }
-
-    /// <summary>
-    /// Realize/recycle handler for the typed TreeView's containers. On realize,
-    /// mounts the node's view into the container's ContentControl; on recycle,
-    /// unmounts it. Does <b>not</b> set <c>args.Handled</c> — the internal
-    /// TreeViewList runs its own ContainerContentChanging handler (indentation /
-    /// selection visuals) and must keep doing so.
-    /// </summary>
-    private void OnTypedTreeContainerContentChanging(
-        WinUI.TreeView treeView, ContainerContentChangingEventArgs args, Action requestRerender)
-    {
-        if (args.ItemContainer?.ContentTemplateRoot is not ContentControl cc) return;
-
-        if (args.InRecycleQueue)
-        {
-            if (cc.Content is UIElement old) UnmountChild(old);
-            cc.Content = null;
-            ClearElementTag(cc);
-            return;
-        }
-
-        PopulateTypedTreeContainer(treeView, args.ItemContainer, args.Item, requestRerender);
-    }
-
-    /// <summary>
-    /// Mounts the node's view into a realized container's ContentControl, unless
-    /// it's already populated. The developer's data item is <c>node.Content</c>.
-    /// </summary>
-    private void PopulateTypedTreeContainer(
-        WinUI.TreeView treeView, WinUI.Control container, object? item, Action requestRerender)
-    {
-        if (GetElementTag(treeView) is not TemplatedTreeViewElementBase el) return;
-        if ((container as ContentControl)?.ContentTemplateRoot is not ContentControl cc) return;
-        if (cc.Content is not null) return; // already hosted for this realization
-        if (item is not WinUI.TreeViewNode node || node.Content is not { } data) return;
-
-        var view = el.BuildView(data);
-        cc.Content = Mount(view, requestRerender);
-        SetElementTag(cc, view);
-    }
-
-    private static void TemplatedTreeItemInvoked(WinUI.TreeView sender, WinUI.TreeViewItemInvokedEventArgs args)
-    {
-        if (args.InvokedItem is WinUI.TreeViewNode node
-            && node.Content is { } item
-            && GetElementTag(sender) is TemplatedTreeViewElementBase el)
-            el.InvokeItemInvoked(item);
-    }
-
-    private static void TemplatedTreeExpanding(WinUI.TreeView sender, WinUI.TreeViewExpandingEventArgs args)
-    {
-        if (args.Node.Content is { } item
-            && GetElementTag(sender) is TemplatedTreeViewElementBase el)
-            el.InvokeExpanding(item);
-    }
-
-    internal UIElement MountTemplatedList(TemplatedListElementBase el, Action requestRerender)
-    {
-        return el.ControlKind switch
-        {
-            TemplatedControlKind.ListView => MountTemplatedListView(el, requestRerender),
-            TemplatedControlKind.GridView => MountTemplatedGridView(el, requestRerender),
-            TemplatedControlKind.FlipView => MountTemplatedFlipView(el, requestRerender),
-            _ => throw new InvalidOperationException($"Unknown TemplatedControlKind: {el.ControlKind}")
-        };
-    }
+    // Mount/Update bodies relocated to Reconciler.TemplatedTree.cs (spec 047 §14).
 
     /// <summary>
     /// Shared ContainerContentChanging handler for all templated items controls.
     /// On materialize: calls viewBuilder, mounts element, stores in ContentControl.
     /// On recycle: unmounts child, clears content.
     /// </summary>
-    private void HandleTemplatedContainerContentChanging(object sender, ContainerContentChangingEventArgs args, Action requestRerender)
+    // Internal so the V1-owned TemplatedListLifecycle can wire it as the shared
+    // realize/recycle handler; also reused 1:1 by Reconciler.KeyedItemsBinding.cs.
+    internal void HandleTemplatedContainerContentChanging(object sender, ContainerContentChangingEventArgs args, Action requestRerender)
     {
         if (args.InRecycleQueue)
         {
@@ -951,184 +503,6 @@ public sealed partial class Reconciler
             // Animation is non-critical — correctness is preserved.
         }
     }
-
-    private WinUI.ListView MountTemplatedListView(TemplatedListElementBase el, Action requestRerender)
-    {
-        var listView = new WinUI.ListView
-        {
-            SelectionMode = el.GetSelectionMode(),
-            IsItemClickEnabled = el.GetIsItemClickEnabled(),
-        };
-        var header = el.GetHeader();
-        if (header is not null) listView.Header = header;
-
-        SetElementTag(listView, el);
-        listView.ItemTemplate = SharedContentControlTemplate.Value;
-
-        listView.ContainerContentChanging += (sender, args) =>
-            HandleTemplatedContainerContentChanging(sender, args, requestRerender);
-
-        listView.SelectionChanged += (s, _) =>
-        {
-            var l = (WinUI.ListView)s!;
-            if (GetElementTag(l) is not TemplatedListElementBase tel) return;
-            tel.InvokeSelectionChanged(l.SelectedIndex);
-            if (tel.HasMultiSelectionCallback)
-            {
-                var snapshot = new List<int>(l.SelectedItems.Count);
-                foreach (var item in l.SelectedItems)
-                {
-                    // Spec 042 Phase 1: items in SelectedItems are
-                    // ReactorRow when the OC delta path is active; preserve
-                    // the legacy int path for the rare direct-int consumer.
-                    if (item is ReactorRow row) snapshot.Add(row.Index);
-                    else if (item is int i) snapshot.Add(i);
-                }
-                tel.InvokeMultiSelectionChanged(snapshot);
-            }
-        };
-        listView.ItemClick += (s, args) =>
-        {
-            var l = (WinUI.ListView)s!;
-            // ClickedItem is the OC element — ReactorRow under the delta
-            // path; int under the legacy path. Translate to the data index
-            // either way.
-            int? idx = args.ClickedItem switch
-            {
-                ReactorRow row => row.Index,
-                int i => i,
-                _ => null,
-            };
-            if (idx is int v)
-                (GetElementTag(l) as TemplatedListElementBase)?.InvokeItemClick(v);
-        };
-
-        // Spec 042 Phase 1: bind to an internally-owned
-        // ObservableCollection<ReactorRow> so insert/remove/move surface
-        // as INotifyCollectionChanged deltas — WinUI animates only the
-        // affected containers rather than re-realizing the entire viewport.
-        var listState = BuildListStateFor(el);
-        SetListState(listView, listState);
-        listView.ItemsSource = listState.Source;
-
-        var selectedIndex = el.GetSelectedIndex();
-        if (selectedIndex >= 0) listView.SelectedIndex = selectedIndex;
-        el.ApplyControlSetters(listView);
-        return listView;
-    }
-
-    /// <summary>
-    /// Builds a fresh <see cref="ReactorListState"/> populated with the
-    /// element's current keys. Used at mount time and at bulk-replace
-    /// bailout time. Tolerates duplicate keys per <see cref="ReactorListState.Reset"/>;
-    /// the bailout path is where duplicate-key diagnostics surface (see
-    /// <see cref="KeyedListDiff"/>).
-    /// </summary>
-    private static ReactorListState BuildListStateFor(TemplatedListElementBase el)
-    {
-        var state = new ReactorListState();
-        int n = el.ItemCount;
-        var seeded = new (int Index, string Key)[n];
-        for (int i = 0; i < n; i++)
-            seeded[i] = (i, el.GetKeyAt(i) ?? $"__null_{i}");
-        state.Reset(seeded);
-        return state;
-    }
-
-    private static ReactorListState BuildListStateForLazy(LazyStackElementBase lazy)
-    {
-        var state = new ReactorListState();
-        int n = lazy.ItemCount;
-        var seeded = new (int Index, string Key)[n];
-        for (int i = 0; i < n; i++)
-            seeded[i] = (i, lazy.GetKeyAt(i) ?? $"__null_{i}");
-        state.Reset(seeded);
-        return state;
-    }
-
-    private WinUI.GridView MountTemplatedGridView(TemplatedListElementBase el, Action requestRerender)
-    {
-        var gridView = new WinUI.GridView
-        {
-            SelectionMode = el.GetSelectionMode(),
-            IsItemClickEnabled = el.GetIsItemClickEnabled(),
-        };
-        var header = el.GetHeader();
-        if (header is not null) gridView.Header = header;
-
-        SetElementTag(gridView, el);
-        gridView.ItemTemplate = SharedContentControlTemplate.Value;
-
-        gridView.ContainerContentChanging += (sender, args) =>
-            HandleTemplatedContainerContentChanging(sender, args, requestRerender);
-
-        gridView.SelectionChanged += (s, _) =>
-        {
-            var g = (WinUI.GridView)s!;
-            if (GetElementTag(g) is not TemplatedListElementBase tel) return;
-            tel.InvokeSelectionChanged(g.SelectedIndex);
-            if (tel.HasMultiSelectionCallback)
-            {
-                var snapshot = new List<int>(g.SelectedItems.Count);
-                foreach (var item in g.SelectedItems)
-                {
-                    if (item is ReactorRow row) snapshot.Add(row.Index);
-                    else if (item is int i) snapshot.Add(i);
-                }
-                tel.InvokeMultiSelectionChanged(snapshot);
-            }
-        };
-        gridView.ItemClick += (s, args) =>
-        {
-            var g = (WinUI.GridView)s!;
-            int? idx = args.ClickedItem switch
-            {
-                ReactorRow row => row.Index,
-                int i => i,
-                _ => null,
-            };
-            if (idx is int v)
-                (GetElementTag(g) as TemplatedListElementBase)?.InvokeItemClick(v);
-        };
-
-        var gridState = BuildListStateFor(el);
-        SetListState(gridView, gridState);
-        gridView.ItemsSource = gridState.Source;
-
-        var selectedIndex = el.GetSelectedIndex();
-        if (selectedIndex >= 0) gridView.SelectedIndex = selectedIndex;
-        el.ApplyControlSetters(gridView);
-        return gridView;
-    }
-
-    private WinUI.FlipView MountTemplatedFlipView(TemplatedListElementBase el, Action requestRerender)
-    {
-        var flipView = new WinUI.FlipView();
-
-        SetElementTag(flipView, el);
-
-        // FlipView doesn't support ContainerContentChanging (not a ListViewBase).
-        // Pre-mount all items — FlipView typically has few items so this is fine.
-        for (int i = 0; i < el.ItemCount; i++)
-        {
-            var itemElement = el.BuildItemView(i);
-            var ctrl = Mount(itemElement, requestRerender);
-            if (ctrl is not null) flipView.Items.Add(ctrl);
-        }
-
-        flipView.SelectionChanged += (s, _) =>
-        {
-            var f = (WinUI.FlipView)s!;
-            (GetElementTag(f) as TemplatedListElementBase)?.InvokeSelectionChanged(f.SelectedIndex);
-        };
-
-        var selectedIndex = el.GetSelectedIndex();
-        if (selectedIndex >= 0) flipView.SelectedIndex = selectedIndex;
-        el.ApplyControlSetters(flipView);
-        return flipView;
-    }
-
-
 
     private static void AddCommandHostAccelerators(WinUI.Grid host, Command[] commands)
     {
@@ -1709,46 +1083,6 @@ public sealed partial class Reconciler
         wrapper.Child = childControl;
         node.RenderedElement = childElement;
         return wrapper;
-    }
-
-    internal UIElement MountLazyStack(LazyStackElementBase lazy, Action requestRerender)
-    {
-        var repeater = new WinUI.ItemsRepeater();
-
-        repeater.Layout = new WinUI.StackLayout
-        {
-            Orientation = lazy.Orientation,
-            Spacing = lazy.Spacing,
-        };
-
-        // Spec 042 Phase 1: bind the repeater to an internally-owned
-        // ObservableCollection<ReactorRow>. Without this, every Items.Count
-        // change replaced the int-range source wholesale and the
-        // ItemsRepeater re-realized every visible child.
-        var listState = BuildListStateForLazy(lazy);
-        SetListState(repeater, listState);
-        repeater.ItemsSource = listState.Source;
-        var factory = lazy.CreateFactory(this, requestRerender, _pool);
-        // Plumb the list state into the factory so its _mountedElements
-        // dictionary is keyed by ReactorRow.Key (reorder-stable) instead
-        // of by realized index.
-        lazy.AttachListStateToFactory(factory, listState);
-        repeater.ItemTemplate = factory;
-        SetElementTag(repeater, lazy);
-        ApplySetters(lazy.RepeaterSetters, repeater);
-
-        var sv = _pool.TryRent(typeof(WinUI.ScrollViewer)) as WinUI.ScrollViewer ?? new WinUI.ScrollViewer();
-        sv.Content = repeater;
-        sv.HorizontalScrollBarVisibility = lazy.Orientation == Microsoft.UI.Xaml.Controls.Orientation.Horizontal
-            ? WinUI.ScrollBarVisibility.Auto
-            : WinUI.ScrollBarVisibility.Disabled;
-        sv.VerticalScrollBarVisibility = lazy.Orientation == Microsoft.UI.Xaml.Controls.Orientation.Vertical
-            ? WinUI.ScrollBarVisibility.Auto
-            : WinUI.ScrollBarVisibility.Disabled;
-        SetElementTag(sv, lazy);
-        ApplySetters(lazy.ScrollViewerSetters, sv);
-
-        return sv;
     }
 
     // ── ItemsView ───────────────────────────────────────────────────────
