+++ src/Reactor/Core/Reconciler.Update.cs
@@ -303,325 +303,9 @@ public sealed partial class Reconciler
     }
 
 
-    internal UIElement? UpdateCheckBox(CheckBoxElement o, CheckBoxElement n, WinUI.CheckBox cb)
-    {
-        SetElementTag(cb, n);
-        bool oldWired = o.OnIsCheckedChanged is not null || o.OnCheckedStateChanged is not null;
-        bool newWired = n.OnIsCheckedChanged is not null || n.OnCheckedStateChanged is not null;
-        if (!oldWired && newWired)
-        {
-            cb.Checked += (s, _) =>
-            {
-                var c = (UIElement)s!;
-                if (ChangeEchoSuppressor.ShouldSuppress(c)) return;
-                var el = GetElementTag(c) as CheckBoxElement;
-                el?.OnIsCheckedChanged?.Invoke(true);
-                el?.OnCheckedStateChanged?.Invoke(true);
-            };
-            cb.Unchecked += (s, _) =>
-            {
-                var c = (UIElement)s!;
-                if (ChangeEchoSuppressor.ShouldSuppress(c)) return;
-                var el = GetElementTag(c) as CheckBoxElement;
-                el?.OnIsCheckedChanged?.Invoke(false);
-                el?.OnCheckedStateChanged?.Invoke(false);
-            };
-            cb.Indeterminate += (s, _) =>
-            {
-                var c = (UIElement)s!;
-                if (ChangeEchoSuppressor.ShouldSuppress(c)) return;
-                var el = GetElementTag(c) as CheckBoxElement;
-                el?.OnCheckedStateChanged?.Invoke(null);
-            };
-        }
-        cb.Content = n.Label;
-        cb.IsThreeState = n.IsThreeState;
-        var target = n.IsThreeState ? n.CheckedState : n.IsChecked;
-        if (cb.IsChecked != target)
-        {
-            ChangeEchoSuppressor.BeginSuppress(cb);
-            cb.IsChecked = target;
-        }
-        ApplySetters(n.Setters, cb);
-        return null;
-    }
 
-    internal UIElement? UpdateExpander(ExpanderElement o, ExpanderElement n, WinUI.Expander exp, Action requestRerender)
-    {
-        exp.IsExpanded = n.IsExpanded;
-        exp.ExpandDirection = n.ExpandDirection;
 
-        // Element header wins over the string slot. Reconcile via ReconcileChild
-        // when both old and new use HeaderTemplate; otherwise swap modes.
-        if (n.HeaderTemplate is not null)
-        {
-            ReconcileChild(o.HeaderTemplate, n.HeaderTemplate,
-                () => exp.Header as UIElement,
-                c => exp.Header = c,
-                () => exp.Header = n.Header,
-                requestRerender);
-        }
-        else
-        {
-            if (exp.Header is UIElement headerCtrl) Unmount(headerCtrl);
-            exp.Header = n.Header;
-        }
 
-        if (!ReferenceEquals(o.ContentTransitions, n.ContentTransitions))
-            exp.ContentTransitions = n.ContentTransitions;
-
-        // Reconcile content child
-        if (exp.Content is UIElement existingContent && CanUpdate(o.Content, n.Content))
-        {
-            var replacement = Update(o.Content, n.Content, existingContent, requestRerender);
-            if (replacement is not null && !ReferenceEquals(exp.Content, replacement))
-                exp.Content = replacement;
-        }
-        else
-        {
-            if (exp.Content is UIElement oldContent)
-                Unmount(oldContent);
-            exp.Content = Mount(n.Content, requestRerender);
-        }
-
-        SetElementTag(exp, n);
-        if (o.OnIsExpandedChanged is null && n.OnIsExpandedChanged is not null)
-        {
-            exp.Expanding += (s, _) => (GetElementTag((UIElement)s!) as ExpanderElement)?.OnIsExpandedChanged?.Invoke(true);
-            exp.Collapsed += (s, _) => (GetElementTag((UIElement)s!) as ExpanderElement)?.OnIsExpandedChanged?.Invoke(false);
-        }
-        ApplySetters(n.Setters, exp);
-        return null;
-    }
-
-    internal UIElement? UpdateNavigationHost(
-        NavigationHostElement oldEl, NavigationHostElement newEl,
-        WinUI.Grid grid, Action requestRerender)
-    {
-        if (!_navigationHostNodes.TryGetValue(grid, out var node))
-        {
-            // Lost tracking — remount from scratch
-            return Mount(newEl, requestRerender);
-        }
-
-        var handle = (Navigation.INavigationHandle)newEl.NavigationHandle;
-        var currentRoute = handle.CurrentRoute;
-
-        // Update the RouteMap if the delegate reference changed (rare but possible if
-        // the parent component recreates the lambda every render).
-        node.RouteMap = newEl.RouteMap;
-
-        // If the handle changed (different navigation stack wired up), re-subscribe
-        if (!ReferenceEquals(node.Handle, handle))
-        {
-            if (node.RouteChangedHandler is not null)
-                node.Handle.RouteChanged -= node.RouteChangedHandler;
-            node.Handle.Detach();
-
-            node.Handle = handle;
-            void onRouteChanged() => requestRerender();
-            handle.RouteChanged += onRouteChanged;
-            node.RouteChangedHandler = onRouteChanged;
-
-            // Re-wire lifecycle guard for the new handle
-            handle.LifecycleGuard = ctx =>
-            {
-                InvokeNavigatingFrom(node.CurrentChildControl, ctx);
-                if (!ctx.IsCancelled)
-                {
-                    node.PendingNavigationMode = ctx.Mode;
-                    node.PendingPreviousRoute = ctx.Route;
-                }
-            };
-        }
-
-        node.RequestRerender = requestRerender;
-
-        if (Equals(currentRoute, node.LastRenderedRoute) && node.CurrentChildElement is not null)
-        {
-            // Route unchanged — reconcile the existing child element in place
-            var newChildElement = node.RouteMap(currentRoute);
-            var replacement = node.CurrentChildControl is not null
-                ? Update(node.CurrentChildElement, newChildElement, node.CurrentChildControl, requestRerender)
-                : Mount(newChildElement, requestRerender);
-
-            if (replacement is not null && node.CurrentChildControl is not null)
-            {
-                // Child control type changed — swap in grid
-                var idx = grid.Children.IndexOf(node.CurrentChildControl);
-                if (idx >= 0)
-                    grid.Children[idx] = replacement;
-                else
-                    grid.Children.Add(replacement);
-                Unmount(node.CurrentChildControl);
-                node.CurrentChildControl = replacement;
-            }
-            else if (replacement is not null)
-            {
-                grid.Children.Add(replacement);
-                node.CurrentChildControl = replacement;
-            }
-
-            node.CurrentChildElement = newChildElement;
-        }
-        else
-        {
-            // Route changed — transition from old page to new page.
-            // Lifecycle sequence per spec:
-            //   1. onNavigatingFrom (already done by LifecycleGuard before stack mutation)
-            //   2-3. Stack mutation (already done)
-            //   4-5. Resolve + mount new element (or restore from cache)
-            //   6. Run transition animation
-            //   7. onNavigatedTo (new page)
-            //   8. onNavigatedFrom (old page)
-            //   9. Unmount or cache old element
-
-            var oldChildControl = node.CurrentChildControl;
-            var oldChildElement = node.CurrentChildElement;
-            var previousRoute = node.LastRenderedRoute;
-            var pendingMode = node.PendingNavigationMode;
-            var pendingPreviousRoute = node.PendingPreviousRoute;
-            node.PendingNavigationMode = null;
-            node.PendingPreviousRoute = null;
-
-            // Collect lifecycle hooks from old page BEFORE detach/unmount
-            var oldHooks = pendingMode is not null
-                ? CollectLifecycleHooks(oldChildControl)
-                : null;
-
-            // Resolve transition: per-navigation override > host default
-            var transitionOverride = handle.PendingTransitionOverride;
-            handle.PendingTransitionOverride = null;
-            var transition = transitionOverride ?? node.HostTransition;
-            var mode = pendingMode ?? Navigation.NavigationMode.Push;
-
-            // Resolve new child: check cache first, then mount fresh
-            UIElement? newChildControl;
-            Element? newChildElement;
-
-            bool wasCacheHit = false;
-            if (node.Cache is not null && node.Cache.TryGet(currentRoute, out var cached))
-            {
-                // Cache hit — restore the mounted control
-                newChildControl = cached.MountedControl;
-                newChildElement = cached.LastElement;
-                node.Cache.Remove(currentRoute);
-                wasCacheHit = true;
-            }
-            else
-            {
-                // Cache miss — mount fresh
-                newChildElement = node.RouteMap(currentRoute);
-                newChildControl = Mount(newChildElement, requestRerender);
-            }
-
-            // Destination-side guard: invoke onNavigatingTo on the new page.
-            // If cancelled, revert to old page.
-            if (!InvokeNavigatingTo(newChildControl, currentRoute, pendingPreviousRoute, mode))
-            {
-                if (!wasCacheHit && newChildControl is not null)
-                    Unmount(newChildControl);
-                return null;
-            }
-
-            // Update node state immediately
-            node.CurrentChildElement = newChildElement;
-            node.CurrentChildControl = newChildControl;
-            node.LastRenderedRoute = currentRoute;
-
-            // Action to finalize the old page (cache or unmount)
-            void FinalizeOldPage(UIElement? oldCtrl, Element? oldElem, object? oldRoute)
-            {
-                if (oldCtrl is null) return;
-                grid.Children.Remove(oldCtrl);
-
-                if (node.Cache is not null && node.CacheMode != Navigation.NavigationCacheMode.Disabled
-                    && oldRoute is not null)
-                {
-                    // Store in cache instead of unmounting
-                    node.Cache.Add(oldRoute, new Navigation.CachedPage
-                    {
-                        MountedControl = oldCtrl,
-                        LastElement = oldElem,
-                        LastAccessed = DateTime.UtcNow,
-                        CacheMode = node.CacheMode,
-                    });
-                }
-                else
-                {
-                    Unmount(oldCtrl);
-                }
-            }
-
-            // Determine whether to run an animated transition
-            bool useAnimation = transition is not Navigation.SuppressTransition
-                && oldChildControl is not null
-                && newChildControl is not null;
-
-            if (useAnimation)
-            {
-                // Mount new content at Opacity 0 alongside old content
-                var inVisual = ElementCompositionPreview.GetElementVisual(newChildControl!);
-                inVisual.Opacity = 0;
-                grid.Children.Add(newChildControl!);
-                node.TransitionInProgress = true;
-
-                // Capture references for the completion callback
-                var capturedOldControl = oldChildControl;
-                var capturedOldElement = oldChildElement;
-                var capturedOldRoute = previousRoute;
-                var capturedNewControl = newChildControl!;
-                var capturedMode = mode;
-                var capturedCurrentRoute = currentRoute;
-                var capturedPreviousRoute = pendingPreviousRoute;
-                var capturedOldHooks = oldHooks;
-
-                Navigation.TransitionEngine.RunTransition(
-                    capturedOldControl!, capturedNewControl, transition, capturedMode,
-                    onComplete: () =>
-                    {
-                        node.TransitionInProgress = false;
-                        FinalizeOldPage(capturedOldControl, capturedOldElement, capturedOldRoute);
-
-                        InvokePostNavigationLifecycle(
-                            capturedNewControl, capturedOldHooks,
-                            capturedCurrentRoute, capturedPreviousRoute, capturedMode);
-                    });
-            }
-            else
-            {
-                // Instant swap (SuppressTransition or missing controls)
-                FinalizeOldPage(oldChildControl, oldChildElement, previousRoute);
-
-                if (newChildControl is not null)
-                    grid.Children.Add(newChildControl);
-
-                InvokePostNavigationLifecycle(
-                    newChildControl, oldHooks,
-                    currentRoute, pendingPreviousRoute, mode);
-            }
-        }
-
-        // Update host properties if changed
-        node.HostTransition = newEl.Transition;
-        if (node.CacheMode != newEl.CacheMode)
-        {
-            node.CacheMode = newEl.CacheMode;
-            if (newEl.CacheMode == Navigation.NavigationCacheMode.Disabled && node.Cache is not null)
-            {
-                node.Cache.Clear();
-                node.Cache = null;
-            }
-            else if (newEl.CacheMode != Navigation.NavigationCacheMode.Disabled && node.Cache is null)
-            {
-                node.Cache = new Navigation.NavigationCache(newEl.CacheSize, evicted => Unmount(evicted));
-            }
-        }
-        if (node.Cache is not null)
-            node.Cache.MaxSize = newEl.CacheSize;
-
-        return null; // Patched in place
-    }
 
     private static object? FindNavItemByTag(global::System.Collections.IEnumerable items, string? selectedTag)
     {
@@ -663,71 +347,6 @@ public sealed partial class Reconciler
     }
 
 
-    // Internal visibility — see comment on MountListView (1.15 Path B).
-    internal UIElement? UpdateListView(ListViewElement o, ListViewElement n, WinUI.ListView lv, Action requestRerender)
-    {
-        lv.SelectionMode = n.SelectionMode;
-        lv.IsItemClickEnabled = n.OnItemClick is not null;
-        if (n.Header is not null) lv.Header = n.Header;
-        if (lv.IncrementalLoadingTrigger != n.IncrementalLoadingTrigger)
-            lv.IncrementalLoadingTrigger = n.IncrementalLoadingTrigger;
-        if (!ReferenceEquals(o.ItemContainerStyle, n.ItemContainerStyle) && n.ItemContainerStyle is not null)
-            lv.ItemContainerStyle = n.ItemContainerStyle;
-
-        // Update ItemsSource — ContainerContentChanging re-mounts visible items via Tag.
-        // Always set a new list when items differ (even same count) so WinUI re-realizes containers.
-        if (!ReferenceEquals(o.Items, n.Items))
-            lv.ItemsSource = Enumerable.Range(0, n.Items.Length).ToList();
-
-        SetElementTag(lv, n);
-
-        // Mount subscribes SelectionChanged unconditionally and reads handlers
-        // via GetElementTag, so no lazy wire here — the tag refresh above
-        // makes a newly-attached OnSelectedIndexChanged / OnSelectionChanged
-        // pick up on the very next selection.
-        if (o.OnItemClick is null && n.OnItemClick is not null)
-            lv.ItemClick += (s, args) =>
-            {
-                var l = (WinUI.ListView)s!;
-                if (args.ClickedItem is int idx)
-                    (GetElementTag(l) as ListViewElement)?.OnItemClick?.Invoke(idx);
-            };
-
-        if (n.SelectedIndex >= 0) lv.SelectedIndex = n.SelectedIndex;
-        ApplySetters(n.Setters, lv);
-        return null;
-    }
-
-    internal UIElement? UpdateGridView(GridViewElement o, GridViewElement n, WinUI.GridView gv, Action requestRerender)
-    {
-        gv.SelectionMode = n.SelectionMode;
-        gv.IsItemClickEnabled = n.OnItemClick is not null;
-        if (n.Header is not null) gv.Header = n.Header;
-        if (gv.IncrementalLoadingTrigger != n.IncrementalLoadingTrigger)
-            gv.IncrementalLoadingTrigger = n.IncrementalLoadingTrigger;
-        if (!ReferenceEquals(o.ItemContainerStyle, n.ItemContainerStyle) && n.ItemContainerStyle is not null)
-            gv.ItemContainerStyle = n.ItemContainerStyle;
-
-        if (!ReferenceEquals(o.Items, n.Items))
-            gv.ItemsSource = Enumerable.Range(0, n.Items.Length).ToList();
-
-        SetElementTag(gv, n);
-
-        // SelectionChanged is wired unconditionally in Mount (see comment in
-        // UpdateListView). Tag refresh suffices to pick up a later-attached
-        // OnSelectedIndexChanged / OnSelectionChanged.
-        if (o.OnItemClick is null && n.OnItemClick is not null)
-            gv.ItemClick += (s, args) =>
-            {
-                var g = (WinUI.GridView)s!;
-                if (args.ClickedItem is int idx)
-                    (GetElementTag(g) as GridViewElement)?.OnItemClick?.Invoke(idx);
-            };
-
-        if (n.SelectedIndex >= 0) gv.SelectedIndex = n.SelectedIndex;
-        ApplySetters(n.Setters, gv);
-        return null;
-    }
 
     /// <summary>
     /// Walks visible (realized) containers and reconciles each item's Element
@@ -739,7 +358,9 @@ public sealed partial class Reconciler
     /// most as null. Children of the realized panel IS the realized set, so
     /// iterating it is O(realized) instead of O(total).
     /// </summary>
-    private void RefreshRealizedContainers(WinUI.ListViewBase listViewBase, Internal.IItemViewSource viewSource, Action requestRerender)
+    // Internal so the V1-owned TemplatedListLifecycle can drive per-row content
+    // reconciliation; also reused 1:1 by Reconciler.KeyedItemsBinding.cs.
+    internal void RefreshRealizedContainers(WinUI.ListViewBase listViewBase, Internal.IItemViewSource viewSource, Action requestRerender)
     {
         var panel = listViewBase.ItemsPanelRoot;
         if (panel is null) return;
@@ -780,176 +401,6 @@ public sealed partial class Reconciler
         }
     }
 
-    internal UIElement? UpdateTemplatedListView(TemplatedListElementBase o, TemplatedListElementBase n, WinUI.ListView lv, Action requestRerender)
-    {
-        lv.SelectionMode = n.GetSelectionMode();
-        lv.IsItemClickEnabled = n.GetIsItemClickEnabled();
-        var header = n.GetHeader();
-        if (header is not null) lv.Header = header;
-
-        // SetElementTag *before* the diff so HandleTemplatedContainerContentChanging
-        // — which fires synchronously inside the OC insert/move ops — reads the
-        // new element when materializing freshly-inserted containers.
-        SetElementTag(lv, n);
-
-        // Spec 042 Phase 1: feed structural changes through the
-        // keyed-diff pipeline so WinUI animates only the affected
-        // containers. RefreshRealizedContainers still runs in every case
-        // because the viewBuilder is a closure that may produce
-        // different content even when the key set is unchanged.
-        ApplyKeyedDiffOrFallback(lv, n);
-        RefreshRealizedContainers(lv, n, requestRerender);
-
-        var selectedIndex = n.GetSelectedIndex();
-        if (selectedIndex >= 0) lv.SelectedIndex = selectedIndex;
-        n.ApplyControlSetters(lv);
-        return null;
-    }
-
-    internal UIElement? UpdateTemplatedGridView(TemplatedListElementBase o, TemplatedListElementBase n, WinUI.GridView gv, Action requestRerender)
-    {
-        gv.SelectionMode = n.GetSelectionMode();
-        gv.IsItemClickEnabled = n.GetIsItemClickEnabled();
-        var header = n.GetHeader();
-        if (header is not null) gv.Header = header;
-
-        SetElementTag(gv, n);
-        ApplyKeyedDiffOrFallback(gv, n);
-        RefreshRealizedContainers(gv, n, requestRerender);
-
-        var selectedIndex = n.GetSelectedIndex();
-        if (selectedIndex >= 0) gv.SelectedIndex = selectedIndex;
-        n.ApplyControlSetters(gv);
-        return null;
-    }
-
-    /// <summary>
-    /// Spec 042 §4 — apply the keyed diff to the control's attached
-    /// <see cref="ReactorListState"/>. If the state is missing (control
-    /// mounted before Phase 1, or the diff bailed out and the legacy
-    /// ItemsSource was swapped in) we transparently rebuild a fresh
-    /// state and re-bind ItemsSource. The result is correctness either
-    /// way; bailout only degrades the animation.
-    /// </summary>
-    private void ApplyKeyedDiffOrFallback(WinUI.ListViewBase lvb, TemplatedListElementBase n)
-    {
-        var state = GetListState(lvb);
-        if (state is null || !ReferenceEquals(lvb.ItemsSource, state.Source))
-        {
-            // No state, or another path replaced ItemsSource (bailout from a
-            // prior render). Rebuild and rebind in one step.
-            var fresh = BuildListStateFromElement(n);
-            SetListState(lvb, fresh);
-            lvb.ItemsSource = fresh.Source;
-            return;
-        }
-
-        // Project new keys through the typed peer. Pass the active ambient
-        // so newly-inserted ReactorRows are tagged with the kind and the
-        // ContainerContentChanging handler can attach a per-container
-        // enter animation as those containers realize. (spec 042 §6.)
-        var ambient = AnimationAmbient.Current;
-        var stats = KeyedListDiff.Apply(
-            state,
-            new TemplatedKeyAdapter(n),
-            static (item, _) => item.Key,
-            _logger,
-            lvb.GetType().Name,
-            ambient,
-            controlInstance: lvb);
-
-        // Drive per-container offset animations for survivors that moved.
-        // Insert / Remove animations attach through the realize/recycle
-        // path so they survive virtualization; Move requires the live
-        // container handle here because the row instance was already
-        // realized before the move op.
-        if (ambient is { HasEffect: true } && stats.MovedRows is { Count: > 0 } movedRows)
-            ApplyMoveAnimations(lvb, movedRows, ambient.Kind);
-
-        if (stats.Bailout)
-        {
-            // Reset replaced state.Source's contents in bulk; ItemsSource
-            // still references the same OC object so WinUI sees a single
-            // Reset action and re-realizes. Acceptable per spec §4.3.
-            // No additional binding refresh needed.
-        }
-    }
-
-    /// <summary>
-    /// Adapter so the generic <see cref="KeyedListDiff.Apply{T}"/> can run
-    /// against the abstract non-generic <see cref="TemplatedListElementBase"/>
-    /// without an extra allocation per item.
-    /// </summary>
-    private readonly struct TemplatedKeyAdapter : IReadOnlyList<TemplatedKeyAdapter.KeyOnly>
-    {
-        private readonly TemplatedListElementBase _el;
-        public TemplatedKeyAdapter(TemplatedListElementBase el) => _el = el;
-        public KeyOnly this[int index] => new(_el.GetKeyAt(index) ?? $"__null_{index}");
-        public int Count => _el.ItemCount;
-        public IEnumerator<KeyOnly> GetEnumerator()
-        {
-            for (int i = 0; i < _el.ItemCount; i++) yield return this[i];
-        }
-        global::System.Collections.IEnumerator global::System.Collections.IEnumerable.GetEnumerator() => GetEnumerator();
-        public readonly record struct KeyOnly(string Key);
-    }
-
-    private static ReactorListState BuildListStateFromElement(TemplatedListElementBase el)
-    {
-        var state = new ReactorListState();
-        int n = el.ItemCount;
-        var seeded = new (int Index, string Key)[n];
-        for (int i = 0; i < n; i++)
-            seeded[i] = (i, el.GetKeyAt(i) ?? $"__null_{i}");
-        state.Reset(seeded);
-        return state;
-    }
-
-    internal UIElement? UpdateTemplatedFlipView(TemplatedListElementBase o, TemplatedListElementBase n, WinUI.FlipView fv, Action requestRerender)
-    {
-        // FlipView items are pre-mounted directly (no ContainerContentChanging).
-        // Build old element array from o, then reconcile like regular items.
-        int oldCount = o.ItemCount;
-        int newCount = n.ItemCount;
-        int shared = Math.Min(oldCount, newCount);
-
-        for (int i = 0; i < shared; i++)
-        {
-            var oldItemElement = o.BuildItemView(i);
-            var newItemElement = n.BuildItemView(i);
-            if (fv.Items[i] is UIElement existingCtrl && CanUpdate(oldItemElement, newItemElement))
-            {
-                var replacement = Update(oldItemElement, newItemElement, existingCtrl, requestRerender);
-                if (replacement is not null && replacement != existingCtrl)
-                    fv.Items[i] = replacement;
-            }
-            else
-            {
-                if (fv.Items[i] is UIElement oldCtrl) Unmount(oldCtrl);
-                fv.Items[i] = Mount(newItemElement, requestRerender)!;
-            }
-        }
-
-        // Remove excess
-        for (int i = oldCount - 1; i >= shared; i--)
-        {
-            if (fv.Items[i] is UIElement oldCtrl) Unmount(oldCtrl);
-            fv.Items.RemoveAt(i);
-        }
-
-        // Add new
-        for (int i = shared; i < newCount; i++)
-        {
-            var ctrl = Mount(n.BuildItemView(i), requestRerender);
-            if (ctrl is not null) fv.Items.Add(ctrl);
-        }
-
-        fv.SelectedIndex = n.GetSelectedIndex();
-        SetElementTag(fv, n);
-        n.ApplyControlSetters(fv);
-        return null;
-    }
-
     private readonly struct ItemsRepeaterKeyAdapter : IReadOnlyList<ItemsRepeaterKeyAdapter.KeyOnly>
     {
         private readonly ItemsRepeaterElementBase _el;
@@ -962,49 +413,6 @@ public sealed partial class Reconciler
         }
         global::System.Collections.IEnumerator global::System.Collections.IEnumerable.GetEnumerator() => GetEnumerator();
         public readonly record struct KeyOnly(string Key);
-    }
-
-    internal UIElement? UpdateLazyStack(LazyStackElementBase n, WinUI.ScrollViewer sv, Action requestRerender)
-    {
-        if (sv.Content is WinUI.ItemsRepeater repeater)
-        {
-            // Try to update the existing factory in place. This avoids
-            // replacing ItemTemplate, which would cause ItemsRepeater to
-            // re-realize all items (modifying Children during layout →
-            // "Cannot run layout in the middle of a collection change").
-            // The factory keeps its identity; existing realized items
-            // stay mounted. On next scroll or layout, IElementFactory.GetElement
-            // uses the updated viewBuilder to produce new content.
-            if (repeater.ItemTemplate is IElementFactory existingFactory && n.TryUpdateFactory(existingFactory))
-            {
-                // Spec 042 Phase 1: route Items changes through the keyed
-                // diff into the internally-owned OC<ReactorRow>. WinUI
-                // sees incremental Insert/Move/RemoveAt events and only
-                // animates affected containers; the steady-state
-                // RefreshRealizedItems below still runs for per-row
-                // content updates.
-                ApplyLazyKeyedDiffOrFallback(repeater, n, existingFactory);
-                n.RefreshRealizedItems(existingFactory, repeater);
-            }
-            else
-            {
-                // First mount or type mismatch — full replacement using the
-                // Phase 1 OC<ReactorRow> binding shape.
-                var fresh = BuildListStateFromLazy(n);
-                SetListState(repeater, fresh);
-                repeater.ItemsSource = fresh.Source;
-                var factory = n.CreateFactory(this, requestRerender, _pool);
-                n.AttachListStateToFactory(factory, fresh);
-                repeater.ItemTemplate = factory;
-            }
-            if (repeater.Layout is WinUI.StackLayout layout)
-                layout.Spacing = n.Spacing;
-            SetElementTag(repeater, n);
-            ApplySetters(n.RepeaterSetters, repeater);
-        }
-        SetElementTag(sv, n);
-        ApplySetters(n.ScrollViewerSetters, sv);
-        return null;
     }
 
     // ── ItemContainer ───────────────────────────────────────────────────
@@ -1029,63 +437,6 @@ public sealed partial class Reconciler
     }
 
 
-    private void ApplyLazyKeyedDiffOrFallback(WinUI.ItemsRepeater repeater, LazyStackElementBase n, IElementFactory factory)
-    {
-        var state = GetListState(repeater);
-        if (state is null || !ReferenceEquals(repeater.ItemsSource, state.Source))
-        {
-            var fresh = BuildListStateFromLazy(n);
-            SetListState(repeater, fresh);
-            repeater.ItemsSource = fresh.Source;
-            n.AttachListStateToFactory(factory, fresh);
-            return;
-        }
-
-        var ambient = AnimationAmbient.Current;
-        var stats = KeyedListDiff.Apply(
-            state,
-            new LazyKeyAdapter(n),
-            static (item, _) => item.Key,
-            _logger,
-            repeater.GetType().Name,
-            ambient,
-            controlInstance: repeater);
-
-        // ItemsRepeater realizes containers through ElementFactory, so the
-        // enter animation runs from there. Moves on already-realized
-        // elements need the same handle-based offset animation as the
-        // templated list path. (spec 042 §6.)
-        if (ambient is { HasEffect: true } && stats.MovedRows is { Count: > 0 } movedRows)
-            ApplyMoveAnimationsRepeater(repeater, movedRows, ambient.Kind);
-        // Bailout reset still mutates state.Source in place, so the
-        // existing ItemsSource binding remains valid.
-    }
-
-    private readonly struct LazyKeyAdapter : IReadOnlyList<LazyKeyAdapter.KeyOnly>
-    {
-        private readonly LazyStackElementBase _el;
-        public LazyKeyAdapter(LazyStackElementBase el) => _el = el;
-        public KeyOnly this[int index] => new(_el.GetKeyAt(index) ?? $"__null_{index}");
-        public int Count => _el.ItemCount;
-        public IEnumerator<KeyOnly> GetEnumerator()
-        {
-            for (int i = 0; i < _el.ItemCount; i++) yield return this[i];
-        }
-        global::System.Collections.IEnumerator global::System.Collections.IEnumerable.GetEnumerator() => GetEnumerator();
-        public readonly record struct KeyOnly(string Key);
-    }
-
-    private static ReactorListState BuildListStateFromLazy(LazyStackElementBase lazy)
-    {
-        var state = new ReactorListState();
-        int n = lazy.ItemCount;
-        var seeded = new (int Index, string Key)[n];
-        for (int i = 0; i < n; i++)
-            seeded[i] = (i, lazy.GetKeyAt(i) ?? $"__null_{i}");
-        state.Reset(seeded);
-        return state;
-    }
-
     /// <summary>
     /// Spec 042 §6 — per-container offset animation for ListView/GridView
     /// survivors that moved index inside an active <see cref="Animations.Animate"/>
@@ -1094,7 +445,9 @@ public sealed partial class Reconciler
     /// ones, which is fine — the realize path attaches the animation when
     /// they come back into view).
     /// </summary>
-    private void ApplyMoveAnimations(WinUI.ListViewBase lvb, IReadOnlyList<ReactorRow> moved, AnimationKind kind)
+    // Internal so the V1-owned TemplatedListLifecycle can drive the keyed-move
+    // offset animation; kept here next to the shared StartMoveOffsetAnimation primitive.
+    internal void ApplyMoveAnimations(WinUI.ListViewBase lvb, IReadOnlyList<ReactorRow> moved, AnimationKind kind)
     {
         var curve = AnimationKindMap.ToCurve(kind);
         if (curve is null) return;
@@ -1134,7 +487,9 @@ public sealed partial class Reconciler
     /// ItemsRepeater doesn't expose <c>ContainerFromItem</c>. Row.Index is
     /// the post-move target position, which is what TryGetElement keys on.
     /// </summary>
-    private void ApplyMoveAnimationsRepeater(WinUI.ItemsRepeater repeater, IReadOnlyList<ReactorRow> moved, AnimationKind kind)
+    // Internal so V1-owned lifecycle classes (LazyStackLifecycle and the templated
+    // families) can reuse the shared repeater move-animation helper.
+    internal void ApplyMoveAnimationsRepeater(WinUI.ItemsRepeater repeater, IReadOnlyList<ReactorRow> moved, AnimationKind kind)
     {
         var curve = AnimationKindMap.ToCurve(kind);
         if (curve is null) return;
@@ -1339,149 +694,7 @@ public sealed partial class Reconciler
 #pragma warning restore CS0618
 
     // ── Typed, data-driven TreeView<T> ───────────────────────────────────
-
-    /// <summary>
-    /// Updates a typed <see cref="TemplatedTreeViewElementBase"/> in place:
-    /// keyed-diffs the node hierarchy and writes back the parent-control props.
-    /// The ItemInvoked / Expanding trampolines resolve the live element via
-    /// <c>GetElementTag</c>, so refreshing the element tag is all that's needed
-    /// to pick up new callbacks — no re-subscription.
-    /// </summary>
-    internal UIElement? UpdateTemplatedTreeView(TemplatedTreeViewElementBase o, TemplatedTreeViewElementBase n, WinUI.TreeView tv, Action requestRerender)
-    {
-        // Diff the node hierarchy (structure + each node's data item). The view
-        // reconcile is a separate flat pass over the realized containers below —
-        // keeping the two concerns decoupled, and using the index-based container
-        // lookup that works under NativeAOT (ContainerFromItem does not resolve
-        // there, and a freshly-realized container can still be the base
-        // ListViewItem rather than TreeViewItem).
-        DiffTemplatedTreeNodes(tv.RootNodes, o, o.GetRoots(), n, n.GetRoots());
-
-        tv.SelectionMode = n.GetSelectionMode();
-        tv.CanDragItems = n.GetCanDragItems();
-        tv.AllowDrop = n.GetAllowDrop();
-        tv.CanReorderItems = n.GetCanReorderItems();
-
-        SetElementTag(tv, n);
-        n.ApplyControlSetters(tv);
-
-        // Reconcile the view of every currently-realized container against its
-        // node's (now-updated) data. Unrealized nodes need no work — their view
-        // is (re)built fresh from node.Content when they next realize via CCC.
-        RefreshRealizedTreeContainers(tv, FindTypedTreeListControl(tv), n, requestRerender);
-        return null;
-    }
-
-    /// <summary>
-    /// Reconciles the hosted view of every realized container against its node's
-    /// current <c>node.Content</c> data. Iterates the flattened
-    /// <see cref="WinUI.ListView.Items"/> via index (the AOT-robust lookup), so
-    /// it covers visible nodes at every depth in one pass.
-    /// </summary>
-    private void RefreshRealizedTreeContainers(WinUI.TreeView tv, WinUI.ListView? list, TemplatedTreeViewElementBase n, Action requestRerender)
-    {
-        if (list is null) return;
-        for (int i = 0; i < list.Items.Count; i++)
-        {
-            if (list.ContainerFromIndex(i) is not ContentControl container) continue;
-            if (container.ContentTemplateRoot is not ContentControl cc) continue;
-            if (list.Items[i] is not WinUI.TreeViewNode node || node.Content is not { } data) continue;
-
-            var newView = n.BuildView(data);
-            if (cc.Content is UIElement existing && GetElementTag(cc) is Element oldView && CanUpdate(oldView, newView))
-            {
-                var replacement = Update(oldView, newView, existing, requestRerender);
-                if (replacement is not null && !ReferenceEquals(cc.Content, replacement))
-                    cc.Content = replacement;
-            }
-            else
-            {
-                if (cc.Content is UIElement old) UnmountChild(old);
-                cc.Content = Mount(newView, requestRerender);
-            }
-            SetElementTag(cc, newView);
-        }
-    }
-
-    private static readonly object[] s_emptyTreeItems = [];
-
-    /// <summary>
-    /// Keyed, in-place hierarchical reconcile of a typed TreeView node list.
-    /// Keys on the element's <c>KeySelector</c>. Matched nodes are reused (their
-    /// data item and any realized container view reconciled in place); the live
-    /// <see cref="WinUI.TreeViewNode"/> collection is then reordered with minimal
-    /// insert/move/remove ops so that <b>unchanged-order updates touch the
-    /// collection not at all</b> — avoiding the container churn that a
-    /// clear-and-rebuild would force.
-    /// </summary>
-    private void DiffTemplatedTreeNodes(
-        IList<WinUI.TreeViewNode> liveNodes,
-        TemplatedTreeViewElementBase oldEl, IReadOnlyList<object> oldItems,
-        TemplatedTreeViewElementBase newEl, IReadOnlyList<object> newItems)
-    {
-        // Snapshot: map old key → (live node, old item). Live nodes correspond
-        // 1:1 to oldItems in order.
-        var oldByKey = new Dictionary<string, (WinUI.TreeViewNode Node, object OldItem)>(oldItems.Count);
-        for (int i = 0; i < oldItems.Count && i < liveNodes.Count; i++)
-            oldByKey.TryAdd(oldEl.GetKey(oldItems[i]), (liveNodes[i], oldItems[i]));
-
-        // Resolve the target node sequence: reuse-and-reconcile matched nodes,
-        // build fresh ones for new keys.
-        var target = new List<WinUI.TreeViewNode>(newItems.Count);
-        for (int i = 0; i < newItems.Count; i++)
-        {
-            var newItem = newItems[i];
-            if (oldByKey.Remove(newEl.GetKey(newItem), out var match))
-            {
-                var node = match.Node;
-                // node.Content is the data item; refresh it so the trampolines
-                // hand back the current T (value-type T is boxed fresh).
-                node.Content = newItem;
-
-                bool expanded = newEl.GetIsExpanded(newItem);
-                if (node.IsExpanded != expanded) node.IsExpanded = expanded;
-
-                DiffTemplatedTreeNodes(
-                    node.Children,
-                    oldEl, oldEl.GetChildren(match.OldItem) ?? s_emptyTreeItems,
-                    newEl, newEl.GetChildren(newItem) ?? s_emptyTreeItems);
-
-                target.Add(node);
-            }
-            else
-            {
-                target.Add(BuildTemplatedTreeNode(newEl, newItem));
-            }
-        }
-
-        // Removed nodes (unmatched old keys): drop them. Their realized
-        // containers recycle → the ContainerContentChanging recycle path tears
-        // their views down.
-        foreach (var leftover in oldByKey.Values)
-            liveNodes.Remove(leftover.Node);
-
-        // Reorder/insert so liveNodes matches `target`. Reused nodes already in
-        // the right slot mean zero collection mutations (the common case).
-        for (int j = 0; j < target.Count; j++)
-        {
-            if (j < liveNodes.Count && ReferenceEquals(liveNodes[j], target[j])) continue;
-
-            int current = IndexOfNode(liveNodes, target[j], j);
-            if (current >= 0) liveNodes.RemoveAt(current);
-            liveNodes.Insert(j, target[j]);
-        }
-        // Trim any stragglers beyond the target length (defensive — removals
-        // above should already have handled these).
-        while (liveNodes.Count > target.Count)
-            liveNodes.RemoveAt(liveNodes.Count - 1);
-    }
-
-    private static int IndexOfNode(IList<WinUI.TreeViewNode> nodes, WinUI.TreeViewNode target, int startAt)
-    {
-        for (int i = startAt; i < nodes.Count; i++)
-            if (ReferenceEquals(nodes[i], target)) return i;
-        return -1;
-    }
+    // Mount/Update bodies relocated to Reconciler.TemplatedTree.cs (spec 047 §14).
 
     private UIElement? UpdateFormField(
         FormFieldElement oldFf, FormFieldElement newFf,
