+++ src/Reactor/Core/V1Protocol/Handlers/NavigationHostHandler.cs
@@ -9,12 +9,10 @@ namespace Microsoft.UI.Reactor.Core.V1Protocol.Handlers;
 /// through V1).
 ///
 /// <para><b>Path B (delegate, no children strategy):</b> delegates Mount /
-/// Update to the engine's existing internal
-/// <see cref="Reconciler.MountNavigationHost"/> /
-/// <see cref="Reconciler.UpdateNavigationHost"/> bodies, which own the
+/// Update to the V1-owned <see cref="NavigationHostLifecycle"/>, which owns the
 /// per-instance route / cache / transition state tracked in the reconciler's
 /// <c>_navigationHostNodes</c> table. <c>Children = null</c> because the
-/// delegate body fully owns child mount/swap inside the host Grid.</para>
+/// lifecycle fully owns child mount/swap inside the host Grid.</para>
 ///
 /// <para><b>Unmount (§4.0.2):</b> teardown is now owned by this handler —
 /// <see cref="Unmount"/> calls <see cref="Reconciler.CleanupNavigationHostNode"/>
@@ -24,7 +22,7 @@ namespace Microsoft.UI.Reactor.Core.V1Protocol.Handlers;
 /// did this in <c>UnmountRecursive</c> is now a V1-OFF-only fallback (deleted
 /// with the V1-OFF escape path in §4.6), so cleanup is byte-identical V1 ON ≡
 /// V1 OFF. The defensive "lost tracking" remount inside
-/// <see cref="Reconciler.UpdateNavigationHost"/> is unreachable under normal V1
+/// <see cref="NavigationHostLifecycle.Update"/> is unreachable under normal V1
 /// operation (the node is always present after <c>Mount</c>), so the void
 /// <see cref="Update"/> here — which cannot substitute the control — preserves
 /// behavior.</para>
@@ -32,10 +30,10 @@ namespace Microsoft.UI.Reactor.Core.V1Protocol.Handlers;
 internal sealed class NavigationHostHandler : IElementHandler<NavigationHostElement, WinUI.Grid>
 {
     public WinUI.Grid Mount(MountContext ctx, NavigationHostElement el)
-        => ctx.Reconciler.MountNavigationHost(el, ctx.RequestRerender);
+        => NavigationHostLifecycle.Mount(ctx.Reconciler, el, ctx.RequestRerender);
 
     public void Update(UpdateContext ctx, NavigationHostElement oldEl, NavigationHostElement newEl, WinUI.Grid ctrl)
-        => ctx.Reconciler.UpdateNavigationHost(oldEl, newEl, ctrl, ctx.RequestRerender);
+        => NavigationHostLifecycle.Update(ctx.Reconciler, oldEl, newEl, ctrl, ctx.RequestRerender);
 
     public void Unmount(UnmountContext ctx, WinUI.Grid control)
         => ctx.Reconciler.CleanupNavigationHostNode(control);
