+++ src/Reactor/Core/V1Protocol/Handlers/TemplatedListHandler.cs
@@ -1,7 +1,4 @@
-using System;
-using System.Diagnostics.CodeAnalysis;
 using Microsoft.UI.Xaml;
-using WinUI = Microsoft.UI.Xaml.Controls;
 
 namespace Microsoft.UI.Reactor.Core.V1Protocol.Handlers;
 
@@ -31,21 +28,10 @@ namespace Microsoft.UI.Reactor.Core.V1Protocol.Handlers;
 internal sealed class TemplatedListHandler : IDecoratorElementHandler<TemplatedListElementBase>
 {
     public UIElement Mount(MountContext ctx, TemplatedListElementBase el)
-        => ctx.Reconciler.MountTemplatedList(el, ctx.RequestRerender);
+        => TemplatedListLifecycle.Mount(ctx.Reconciler, el, ctx.RequestRerender);
 
     public UIElement Update(UpdateContext ctx, TemplatedListElementBase oldEl, TemplatedListElementBase newEl, UIElement control)
-    {
-        var rr = ctx.RequestRerender;
-        UIElement? replacement = control switch
-        {
-            WinUI.GridView gv => ctx.Reconciler.UpdateTemplatedGridView(oldEl, newEl, gv, rr),
-            WinUI.FlipView fv => ctx.Reconciler.UpdateTemplatedFlipView(oldEl, newEl, fv, rr),
-            WinUI.ListView lv => ctx.Reconciler.UpdateTemplatedListView(oldEl, newEl, lv, rr),
-            _ => throw new InvalidOperationException(
-                $"TemplatedListHandler: unexpected control type {control.GetType().Name} for element {newEl.GetType().Name}."),
-        };
-        return replacement ?? control;
-    }
+        => TemplatedListLifecycle.Update(ctx.Reconciler, oldEl, newEl, control, ctx.RequestRerender) ?? control;
 
     public V1UnmountDisposition Unmount(UnmountContext ctx, TemplatedListElementBase? element, UIElement control)
         => V1UnmountDisposition.ContinueDefaultTraversal;
