+++ tests/Reactor.AppTests.Host/SelfTest/SelfTestFixtureRegistry.cs
@@ -220,6 +220,7 @@ internal static class SelfTestFixtureRegistry
         "Navigation_NavHostGoBack",
         "Navigation_NavHostLifecycleOrder",
         "Navigation_NavHostNested",
+        "Navigation_NavHostCacheMode",
         "Observable_UseObservable_Rerender",
         "Observable_UseObservable_ExternalMutation",
         "Observable_UseObservableProperty_FineGrained",
@@ -1385,6 +1386,7 @@ internal static class SelfTestFixtureRegistry
         "Navigation_NavHostGoBack" => new NavigationFixtures.NavHostGoBack(harness),
         "Navigation_NavHostLifecycleOrder" => new NavigationFixtures.NavHostLifecycleOrder(harness),
         "Navigation_NavHostNested" => new NavigationFixtures.NavHostNested(harness),
+        "Navigation_NavHostCacheMode" => new NavigationFixtures.NavHostCacheMode(harness),
         "Observable_UseObservable_Rerender" => new ObservableFixtures.UseObservable_Rerender(harness),
         "Observable_UseObservable_ExternalMutation" => new ObservableFixtures.UseObservable_ExternalMutation(harness),
         "Observable_UseObservableProperty_FineGrained" => new ObservableFixtures.UseObservableProperty_FineGrained(harness),
